+++ super_editor/lib/src/super_reader/read_only_document_android_touch_interactor.dart
@@ -16,8 +16,7 @@ import 'package:super_editor/src/infrastructure/platforms/android/android_docume
 import 'package:super_editor/src/infrastructure/platforms/mobile_documents.dart';
 import 'package:super_editor/src/infrastructure/selection_leader_document_layer.dart';
 import 'package:super_editor/src/infrastructure/touch_controls.dart';
-
-import '../infrastructure/super_textfield/metrics.dart';
+import 'package:super_editor/src/super_textfield/metrics.dart';
 
 /// Read-only document gesture interactor that's designed for Android touch input, e.g.,
 /// drag to scroll, and handles to control selection.
