+++ super_editor/lib/src/super_textfield/android/_caret.dart
