+++ super_editor/lib/src/super_textfield/infrastructure/outer_box_shadow.dart
