+++ super_editor/lib/src/super_textfield/ios/ios_textfield.dart
@@ -6,22 +6,20 @@ import 'package:super_editor/src/infrastructure/attributed_text_styles.dart';
 import 'package:super_editor/src/infrastructure/flutter/flutter_pipeline.dart';
 import 'package:super_editor/src/infrastructure/focus.dart';
 import 'package:super_editor/src/infrastructure/ime_input_owner.dart';
+import 'package:super_editor/src/infrastructure/platforms/ios/toolbar.dart';
 import 'package:super_editor/src/infrastructure/platforms/mobile_documents.dart';
-import 'package:super_editor/src/infrastructure/super_textfield/infrastructure/fill_width_if_constrained.dart';
-import 'package:super_editor/src/infrastructure/super_textfield/infrastructure/hint_text.dart';
-import 'package:super_editor/src/infrastructure/super_textfield/infrastructure/text_scrollview.dart';
-import 'package:super_editor/src/infrastructure/super_textfield/input_method_engine/_ime_text_editing_controller.dart';
-import 'package:super_editor/src/infrastructure/super_textfield/ios/_editing_controls.dart';
+import 'package:super_editor/src/super_textfield/infrastructure/fill_width_if_constrained.dart';
+import 'package:super_editor/src/super_textfield/infrastructure/hint_text.dart';
+import 'package:super_editor/src/super_textfield/infrastructure/text_scrollview.dart';
+import 'package:super_editor/src/super_textfield/input_method_engine/_ime_text_editing_controller.dart';
+import 'package:super_editor/src/super_textfield/ios/_editing_controls.dart';
 import 'package:super_text_layout/super_text_layout.dart';
 
-import '../../platforms/ios/toolbar.dart';
 import '../metrics.dart';
 import '../styles.dart';
 import '_floating_cursor.dart';
 import '_user_interaction.dart';
 
-export '../../platforms/ios/selection_handles.dart';
-export '../../platforms/ios/toolbar.dart';
 export '../infrastructure/magnifier.dart';
 export '_caret.dart';
 export '_user_interaction.dart';
