+++ super_editor/lib/super_editor.dart
@@ -64,13 +64,15 @@ export 'src/infrastructure/keyboard.dart';
 export 'src/infrastructure/multi_tap_gesture.dart';
 export 'src/infrastructure/pausable_value_notifier.dart';
 export 'src/infrastructure/platforms/android/android_document_controls.dart';
+export 'src/infrastructure/platforms/android/toolbar.dart';
 export 'src/infrastructure/platforms/ios/ios_document_controls.dart';
+export 'src/infrastructure/platforms/ios/toolbar.dart';
 export 'src/infrastructure/platforms/mobile_documents.dart';
 export 'src/infrastructure/scrolling_diagnostics/scrolling_diagnostics.dart';
 export 'src/infrastructure/selection_leader_document_layer.dart';
 export 'src/infrastructure/signal_notifier.dart';
 export 'src/infrastructure/strings.dart';
-export 'src/infrastructure/super_textfield/super_textfield.dart';
+export 'src/super_textfield/super_textfield.dart';
 export 'src/infrastructure/touch_controls.dart';
 export 'src/infrastructure/text_input.dart';
 export 'src/infrastructure/viewport_size_reporting.dart';
