+++ super_editor/lib/super_text_field.dart
@@ -1,0 +1,3 @@
+library super_text_field;
+
+export 'src/super_textfield/super_textfield.dart';
